refactor: reorganization of json models for clearer structure
