+++ src/SmartHopper.Core.Grasshopper/Utils/GHObjectFactory.cs
@@ -8,10 +8,10 @@
  * version 3 of the License, or (at your option) any later version.
  */
 
-using Grasshopper;
-using Grasshopper.Kernel;
 using System;
 using System.Diagnostics;
+using Grasshopper;
+using Grasshopper.Kernel;
 
 namespace SmartHopper.Core.Grasshopper.Utils
 {
