+++ src/SmartHopper.Core/Models/Components/ComponentProperty.cs
@@ -1,0 +1,39 @@
+/*
+ * SmartHopper - AI-powered Grasshopper Plugin
+ * Copyright (C) 2024 Marc Roca Musach
+ * 
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ */
+
+using System;
+using Newtonsoft.Json;
+
+namespace SmartHopper.Core.Models.Components
+{
+    /// <summary>
+    /// Represents a property of a Grasshopper component with type information and human-readable format.
+    /// </summary>
+    public class ComponentProperty
+    {
+        /// <summary>
+        /// The actual value of the property.
+        /// </summary>
+        [JsonProperty("value")]
+        public object Value { get; set; }
+
+        /// <summary>
+        /// The type name of the property.
+        /// </summary>
+        [JsonProperty("type")]
+        public string Type { get; set; }
+
+        /// <summary>
+        /// A human-readable representation of the property value.
+        /// </summary>
+        [JsonProperty("humanReadable")]
+        public string HumanReadable { get; set; }
+    }
+}
